requirements.txt
pyyaml>=6.0
torch>=2.5.1
pyaudio>=0.2.14
numpy>=2.2.0
soundfile>=0.13.0
pytest>=7.4.0
torchaudio>=2.5.1
openai-whisper>=20240930
scipy>=1.11.0
psutil>=5.9.0